Gestion de errores artista

+++ src/main/java/com/salesianostriana/dam/trianafy/service/ArtistService.java
@@ -1,9 +1,11 @@
 package com.salesianostriana.dam.trianafy.service;
 
 
+import com.salesianostriana.dam.trianafy.exception.ArtistNotFoundException;
 import com.salesianostriana.dam.trianafy.model.Artist;
 import com.salesianostriana.dam.trianafy.repos.ArtistRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -17,27 +19,44 @@ public class ArtistService {
 
 
     public Artist add(Artist artist) {
+
         return repository.save(artist);
     }
 
-    public Optional<Artist> findById(Long id) {
-        return repository.findById(id);
+    public Artist findById(Long id) {
+
+        return repository.findById(id)
+                .orElseThrow(() -> new ArtistNotFoundException(id));
     }
 
     public List<Artist> findAll() {
-        return repository.findAll();
+
+        List<Artist> result = repository.findAll();
+
+        if(result.isEmpty()){
+            throw new ArtistNotFoundException();
+        }
+
+        return result;
     }
 
-    public Artist edit(Artist artist) {
-        return repository.save(artist);
+    public Artist edit(Long id, Artist artist) {
+        return repository.findById(id)
+                .map(old ->{
+                    old.setName(artist.getName());
+                    return repository.save(old);
+                }).orElseThrow(() -> new ArtistNotFoundException(id));
     }
 
     public void delete(Artist artist) {
+        if (repository.existsById(artist.getId()))
         repository.delete(artist);
     }
 
     public void deleteById(Long id) {
-        repository.deleteById(id);
+
+        if (repository.existsById(id))
+            repository.deleteById(id);
     }
 
 }
